FPS tutorial
GameMode - the basic class that defines the rules and conditions of the game.
Skeleta Mesh - Mesh with animation.
If you remove the third-person Mesh shown in the character's constructor stage, it does not apply when you play the game. It does not take effect until you do so in BeginPlay.