FPS tutorial 2
In order to use other classes, two types are required: include and TSubclassOf.
Most of the functions required for the game are implemented on a component basis, so use them properly.
Collision Generation Channel -> Presetting allows custom collisions to be created and requires the BodyInstance name to be changed to the corresponding collisions at the constructor stage of the class to be used (most of the collisions are set up at the constructor stage). )
Can create a custom HUD.