casting
Most objects are created by inheriting the AActor class.
The same is true of projectiles and objects in FPS games, and to detect collisions through collisions, AActor classes are cast and used in the implementation of NotifyActorBeginOverlap function, which can be used in BP as well (only BP was used in the last implementation).